+++ devknife/__init__.py
@@ -1,5 +1,5 @@
 """
-Python DevKnife Toolkit - 개발자를 위한 올인원 터미널 유틸리티 툴킷
+Nalutbae DevKnife Toolkit - 개발자를 위한 올인원 터미널 유틸리티 툴킷
 
 이 패키지는 개발자들이 일상적으로 사용하는 다양한 유틸리티 기능들을
 하나의 통합된 도구로 제공합니다.